requirements.txt
scipy>=1.8.1

torch~=2.4.0+cu124
python-dotenv~=1.0.1
llama-index-core~=0.11.9
transformers~=4.43.4
pillow~=10.4.0